backend/requirements.txt
# FastAPI + Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
geoalchemy2==0.15.2
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ML / Vision
ultralytics==8.3.50
opencv-python-headless==4.10.0.84
Pillow==11.0.0
numpy==1.26.4

# Hotspot prediction
xgboost==2.1.3
scikit-learn==1.6.0
pandas==2.2.3
joblib==1.4.2

# HTTP client (weather API)
httpx==0.28.1

# Push notifications
exponent-server-sdk==2.1.0

# Settings
pydantic-settings==2.7.0

# Utils
python-dateutil==2.9.0
aiofiles==24.1.0
